% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BesselFunctions.r
\name{BesselIR}
\alias{BesselIR}
\title{computes the ratio of two BesselI frunctions.}
\usage{
BesselIR(x, nu, ifLog = FALSE)
}
\arguments{
\item{x:}{A positive real number or a vector of real numbers. The value at which the Ratio of Bessel to be calculated.}

\item{nu:}{A positive real number, the order of BesselI}

\item{ifLog:}{if ifLog=TRUE  is specified then log of the Bessel function will be retunred}
}
\value{
BesselI evaluated at x>0.
\deqn{ \frac{ I_{nu+1}(x) }{ I_{nu}(x) }  }.
}
\description{
computes the ratio of two BesselI frunctions.
}
\examples{
BesselIR(x=10, nu=6,  ifLog=TRUE)
}
