% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BesselFunctions.r
\name{BesselIRatio}
\alias{BesselIRatio}
\title{computes the ratio of two BesselI frunctions.}
\usage{
BesselIRatio(x, nuNum, nuDeno, ifLog = FALSE)
}
\arguments{
\item{x:}{A positive real number or a vector of real numbers. The value at which the Ratio of Bessel to be calculated.}

\item{nuNum:}{A positive real number, the order of BesselI in the numerator}

\item{nuDeno:}{A positive real number, the order of BesselI in the denominator}

\item{ifLog:}{if ifLog=TRUE  is specified then log of the ratio of the Bessel function will be retunred
Number of Terms added in the series represantion of the density.}
}
\value{
Ratio of Bessel function
\deqn{   \frac{ I_{nuNum}(x) }{ I_{nuDeno}(x) }  }
}
\description{
computes the ratio of two BesselI frunctions.
}
\examples{
BesselIRatio(x=10, nuNum=5,nuDeno=7,  ifLog=TRUE)
BesselIRatio(x=10000000000, nuNum=5, nuDeno=7,  ifLog=TRUE)
}
