% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Approx_MPG_GIG.r
\name{CalculateTV_between_MPG_GIG_CDF}
\alias{CalculateTV_between_MPG_GIG_CDF}
\title{CalculateTV_between_MPG_GIG_CDF}
\usage{
CalculateTV_between_MPG_GIG_CDF(
  alpha,
  nu,
  search_n = 100,
  l_lim = NULL,
  r_lim = NULL
)
}
\arguments{
\item{alpha:}{a positive real number}

\item{nu:}{A positive real number of the Ratio bessel \deqn{\frac{I_{\nu+1}(\alpha)}{I_{\nu}(\alpha)}}}
}
\value{
Total Variation Distribution between MPG(alpha, nu) and the fowwowing distributions:
GIG(lambda=-(nu+1), .5, 2\alpha^2), GIG(lambda=\nu_{opt}, .5, 2\alpha^2)
}
\description{
CalculateTV_between_MPG_GIG_CDF
}
\examples{
alpha=1
CalculateTV_between_MPG_GIG_CDF(alpha=10, nu=1 )
}
