% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare_MPG_GIG_Characteristic_functions.r
\name{comp_characteristic_function_alt}
\alias{comp_characteristic_function_alt}
\title{comp_characteristic_function_alt}
\usage{
comp_characteristic_function_alt(
  t,
  fun = NULL,
  Name = NULL,
  showBcGround = TRUE,
  colr = c("white", "black")
)
}
\arguments{
\item{t}{=c( -20000:20000);}

\item{alpha}{=10;}

\item{nu}{=100}
}
\description{
comp_characteristic_function_alt
}
\examples{
t=c( -20000:20000);  alpha=10; nu=100
comp_characteristic_function_alt(t,fun = c("ch_GIG", "ch_MPG"), c("GIG ","MPG"), showBcGround=FALSE, colr = c("red","blue") )
comp_characteristic_function_alt(t,fun = c("ch_GIG", "ch_MPG"), c("GIG ","MPG"), showBcGround=TRUE )
}
