% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare_MPG_GIG_Characteristic_functions.r
\name{comp_characteristic_functions_2d}
\alias{comp_characteristic_functions_2d}
\title{comp_characteristic_functions_2d}
\usage{
comp_characteristic_functions_2d(t, fun = NULL, Name = NULL)
}
\description{
comp_characteristic_functions_2d
}
\examples{
t=c( -20000:20000);  alpha=10; nu=100
aa=comp_characteristic_functions_2d(t,fun = c("ch_GIG", "ch_MPG"), c("GIG ","MPG") )
plot(aa)
t=c( -20000:20000)/80;  alpha=5; nu=0
aa=comp_characteristic_functions_2d(t,fun = c("ch_GIG", "ch_MPG"), c("GIG ","MPG") )
plot(aa)
}
