% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMPG.r
\name{dMPG}
\alias{dMPG}
\title{computes the density frunction for the Modfified Ploya Gamma (MPG) distribution.}
\usage{
dMPG(x, alpha, nu, iflog = FALSE, Num_of_Terms = 200)
}
\arguments{
\item{x:}{A positive real number or a vector of real numbers. The value at which the density to be calculated.}

\item{alpha:}{A positive real number.}

\item{nu:}{A positive real number, order of the MPG distribution.}

\item{Num_of_Terms:}{Integer greater than 50.
Number of Terms added in the series represantion of the density.}
}
\value{
Density of the MPG(alpha, nu) evaluated at x
}
\description{
computes the density frunction for the Modfified Ploya Gamma (MPG) distribution.
}
\examples{
library(Bessel)
dMPG(.05, alpha=5, nu=0)
dMPG(c(.01, .05,.1) , alpha=5, nu=0)
plot(function(x){dMPG(x, alpha=5, nu=0, iflog = FALSE)}, xlim=c(.005, .2))
}
