% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_functions.R
\name{KAPPA_INITIAL}
\alias{KAPPA_INITIAL}
\title{Calculates initial value of KAPPA_INITIAL(Y). Approximate posterior mode.}
\usage{
KAPPA_INITIAL(Y)
}
\arguments{
\item{Y:}{n\times d  matrix containing n directional data of dimension d.
norm of each row of the matrix Y is 1.}
}
\value{
Approximate posterior mode for the concentration parameter.
}
\description{
Calculates initial value of KAPPA_INITIAL(Y). Approximate posterior mode.
}
\examples{
library(Rfast)
Y=rvmf(n =10, mu=c(1,0,0),k = 10)
KAPPA_INITIAL(Y)
}
