% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Approx_MPG_GIG.r
\name{matching_mean_GIG_nu_calculator}
\alias{matching_mean_GIG_nu_calculator}
\title{matching_mean_GIG_nu_calculator}
\usage{
matching_mean_GIG_nu_calculator(alpha, nu)
}
\arguments{
\item{alpha:}{a positive real number}

\item{nu:}{A positive real number of the Ratio bessel \deqn{\frac{I_{\nu+1}(\alpha)}{I_{\nu}(\alpha)}}}
}
\value{
opt_nu_for_GIG: An uniroot object for the optimum. opt_nu_for_GIG$root is the optimum nu_1 that satisfies the equation $\frac{K_{\nu_1+1}(\alpha)}{K_{\nu_1}(\alpha)}= \frac{I_{\nu+1}(\alpha)}{I_{\nu}(\alpha)}$
}
\description{
matching_mean_GIG_nu_calculator
}
\examples{
alpha=1
matching_nu_calculator(alpha=10, nu=1 )
}
