% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_BLASSO_Dir_regression_sampler_V1.R
\name{MCMC_BLASSO_Dir_regression_sampler_V1}
\alias{MCMC_BLASSO_Dir_regression_sampler_V1}
\title{Performs a Bayesian regression for the directional responses.}
\usage{
MCMC_BLASSO_Dir_regression_sampler_V1(
  Y,
  X,
  prior = NULL,
  beta_init = NULL,
  Sigma_init = NULL,
  MCSamplerSize = 50,
  K = 100,
  eps_accuracy = 1e-08,
  lasso_lambda = 0.01,
  Sample_lasso_lambda = NULL,
  n_less_than_p_sampler = "JOHNDROW",
  lasso_lambda_spec = list(Type = "SAMPLE", lasso_lambda = 0.01, hyper_lambda_selector =
    list(Lambda_lower = 0.01, Lambda_upper = 0.02, info_weight = 10))
)
}
\arguments{
\item{Y:}{n times d  matrix containing n directional response of dimension parameter d.
norm of each row of the matrix Y is 1.}

\item{X:}{A n*p design matrix containing the data of the p variables.
If it is required, the intercept should be included in the first column.}

\item{prior:}{Prior specification for the posterior. Default (prior=NULL) uses the very flat prior.}

\item{beta_init:}{Initial value of rregression coefficient. The default (beta_innit=NULL) utilizes the EM algorithm to get a initial value}

\item{MCSamplerSize:}{Number of MCMC samples required to be generated.}

\item{n_less_than_p_sampler:}{Two choices 'JOHNDROW' or STANDARD'. The option 'JOHNDROW' is more efficient when n>>p. for example n=80 p =200}

\item{lasso_lambda:}{lasso_lambda=.01 is the default value for the Lasso tuning parameter lambda.
If we want to use a fixed lambda select additionally lasso_lambda_spec=list(Type="FIXED")}

\item{lasso_lambda_spec:}{lasso_lambda_spec=list(Type="SAMPLE",lasso_lambda=.01,hyper_lambda_selector= list(Lambda_lower=0.01, Lambda_upper=0.02,info_weight=10 ) ),
if Type="SAMPLE" is selected then posterior samples for lambda parameter (lasso tuning parameter) is  drawn from the corresponding posterior.
If Type="Fixed" it will take a fixed value of lambda:  lasso_lambda=.01
It is recommended to use the Type="SAMPLE" specification.}
}
\value{
MCMC samples from for  Posterior of mode and concentration parameter of Vonmises distribution.
}
\description{
Performs a Bayesian regression for the directional responses.
}
\examples{
library(Rfast)
#library(benchmarkme)
library(Bessel)
library(gsl)
library(ghyp)
n=1000 # NUmber of the samples
p=4  # NUmber of the regression covariates
d=5 # Number of direcions in the direcional data
data_lst = Data_generator_vnf_reg(n=n, p=p, d=d, concentration_factor = 1, beta_factor = 5)
Y = data_lst$Y;X=data_lst$X;
lst=MCMC_BLASSO_Dir_regression_sampler_V1(Y=Y, X=X,  MCSamplerSize =100)
i=1;j= 1
library(ggplot2)
library(cowplot) #
Plot_MCMC_Diag_Triplet(lst$MC$Mc_Beta[,i,j],y_lab_text = bquote(beta[.(i)][.(j)]))
}
