% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPGMeanSd.R
\name{MPG_mean_sd}
\alias{MPG_mean_sd}
\title{Generates samples from Modfified Ploya Gamma (MPG) distribution.}
\usage{
MPG_mean_sd(alpha = 1, nu)
}
\arguments{
\item{alpha:}{A positive real number. It is a parameter of the distribution.}

\item{nu:}{A positive real number, order of the MPG distribution.}
}
\value{
Mean (mean) and Standard Deviation (sd) of the MPG(alpha, nu) distribution.
}
\description{
Generates samples from Modfified Ploya Gamma (MPG) distribution.
}
\examples{
library(Bessel)
MPG_mean_sd(alpha=5, nu=0)
data=rMPG(n=1000, alpha=5, nu=0)
#' data=rMPG(n=1000, alpha=50, nu=0)
mean(data)
sd(data)
}
