% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin_plot_ACF.r
\name{plot_Acf_violiin}
\alias{plot_Acf_violiin}
\title{plot_Acf_violiin_Slice}
\usage{
plot_Acf_violiin(
  data,
  MAXLAG = 15,
  scale_type = "width",
  size = 0.5,
  Ylimit = NULL,
  ifGray = TRUE,
  GridLines = "Vertical"
)
}
\arguments{
\item{data}{:  A dataframe of size n*lag X number of different kappa seq.}
}
\description{
plot_Acf_violiin_Slice
}
\examples{
data(ACF_Slice)
plot_Acf_violiin(ACF_Slice, Ylimit=c(.95, 1))
data(ACF_Sim1_3D_MPG_DA_vioplot_data)
plot_Acf_violiin(Vio_ACF_MPG, size=.2, ifGray=TRUE)
data(ACF_Sim1_3D_GEO_DA_vioplot_data)
plot_Acf_violiin(Vio_ACF_MPG, size=.2, ifGray=TRUE, GridLines="Vertical")
data(ACF_sim2_pD_MPG_DA_vioplot_data)
plot_Acf_violiin(Vio_ACF_MPG_pd,size=.2, ifGray=TRUE, GridLines="none")
}
