% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin_plot_ACF.r
\name{plot_Acf_violiin_Slice}
\alias{plot_Acf_violiin_Slice}
\title{plot_Acf_violiin_Slice}
\usage{
plot_Acf_violiin_Slice(data, MAXLAG = 15, scale_type = "width")
}
\arguments{
\item{data}{:  A dataframe of size n*lag X number of different kappa seq.}
}
\description{
plot_Acf_violiin_Slice
}
\examples{
data(ACF_Slice)
plot_Acf_violiin_Slice(ACF_Slice)
}
