% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Approx_MPG_GIG.r
\name{Plot_Approximation_MPG_GIG_Relative_diff}
\alias{Plot_Approximation_MPG_GIG_Relative_diff}
\title{Plot_Approximation_MPG_GIG_Relative_diff}
\usage{
Plot_Approximation_MPG_GIG_Relative_diff(x, alpha, nu = 0, bgCol = "#686868")
}
\description{
Plot_Approximation_MPG_GIG_Relative_diff
}
\examples{
x=seq(from = .01,to=.15,by = .0001)
alpha=seq(from=10, to = 30, by = 1)
pp=Plot_Approximation_MPG_GIG_Relative_diff(x=seq(from = .01,to=.15,by = .001),alpha= seq(from=10, to = 20, by = 1))
pp$p
pp$p1
}
