% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMPG_nu_Half.R
\name{plot_GIG_MPG_Proposal_h_densities}
\alias{plot_GIG_MPG_Proposal_h_densities}
\title{Plots the density and the proposal density}
\usage{
plot_GIG_MPG_Proposal_h_densities(alpha = 5, UpperLim = NULL, LowerLim = NULL)
}
\description{
Plots the density and the proposal density
}
\examples{
plot_GIG_MPG_Proposal_h_densities(3)
plot_GIG_MPG_Proposal_h_densities(5)
}
