% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Diagonistics_Triplet_Plot.R
\name{Plot_MCMC_Diag_Triplet}
\alias{Plot_MCMC_Diag_Triplet}
\title{Plotting ACF, Density, Trace and Cumulative Average plots.}
\usage{
Plot_MCMC_Diag_Triplet(
  data,
  lag.max = 24,
  ci = 0.95,
  large.sample.size = TRUE,
  horizontal = TRUE,
  CumType = "both",
  ThemeBoth = 2,
  ACF = TRUE,
  Density = TRUE,
  y_lab_text = NULL,
  SelBw = "nrd0",
  ...
)
}
\description{
Plotting ACF, Density, Trace and Cumulative Average plots.
}
\examples{
  y_lab_text=expression(paste('Cumulative Average of ',Var_name  ))
  y_lab_text=expression(beta^T*beta)

  Plot_post_sample_cumAvg_Trace_Acf(data = kappa_all[,12],horizontal = FALSE, ACF = TRUE, Density=TRUE, y_lab_text = expression(paste("Posterior Samples of ", kappa)))
Plot_post_sample_cumAvg_Trace_Acf
}
