% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Approx_MPG_GIG.r
\name{Plot_MPG_GIG_CDF}
\alias{Plot_MPG_GIG_CDF}
\title{Plot_MPG_GIG_CDF}
\usage{
Plot_MPG_GIG_CDF(alpha, nu, gigCompare = TRUE, l_lim = NULL, r_lim = NULL)
}
\description{
Plot_MPG_GIG_CDF
}
\examples{
alpha=1
pp=Plot_MPG_GIG_CDF(alpha= alpha, nu=1, gigCompare=TRUE )
}
