% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Approx_MPG_GIG.r
\name{Plot_MPG_GIG_CDF_alt}
\alias{Plot_MPG_GIG_CDF_alt}
\title{Plot_MPG_GIG_CDF}
\usage{
Plot_MPG_GIG_CDF_alt(alpha, nu, gigCompare = TRUE, l_lim = NULL, r_lim = NULL)
}
\arguments{
\item{alpha:}{concentration parameter of the MPG distribution. A positive real number}

\item{nu:}{order of the MPG distribution. A positive real number}
}
\description{
Plot_MPG_GIG_CDF
}
\examples{
alpha=1
pp=Plot_MPG_GIG_CDF_alt(alpha= alpha, nu=1, gigCompare=TRUE )
}
