% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IEHS_sampler.r
\name{rIEHC}
\alias{rIEHC}
\title{Generates samples from Inverse Expontially Tilted Half Cauchy (IEHC) distribution.}
\usage{
rIEHC(n = 1, a, b)
}
\arguments{
\item{n:}{Number of samples to be generated.}

\item{a:}{A positive real number, IEHCshape parameter in latex.}

\item{b:}{A positive real number. IEHCscale parameter in latex.}
}
\value{
Samples from the IEHC(n, a, b) distribution.
}
\description{
Generates samples from Inverse Expontially Tilted Half Cauchy (IEHC) distribution.
}
\examples{
data=rIEHC(n=1000, a=1, b=.1)
}
