% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMatrixNormal.R
\name{rMatrixNormal}
\alias{rMatrixNormal}
\title{rMatrixNormal(M,V1,V2)
This function generates single sample from a matrix normal random variable}
\usage{
rMatrixNormal(M, V1, V2)
}
\arguments{
\item{$M$}{the matrix which is the mean of the matrix Normal random variable}

\item{$V1$}{and $V2$ are variance component such that the variance of Vec(X) is V2 cross V1}
}
\description{
rMatrixNormal(M,V1,V2)
This function generates single sample from a matrix normal random variable
}
