% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMPG.r
\name{rMPG}
\alias{rMPG}
\title{Generates samples from Modfified Ploya Gamma (MPG) distribution.}
\usage{
rMPG(n = 1, alpha, nu, Num_of_Terms = 200, eps_accuracy = 1e-05)
}
\arguments{
\item{n:}{Number of samples to be generated.}

\item{alpha:}{A positive real number.}

\item{nu:}{A positive real number, order of the MPG distribution.}

\item{Num_of_Terms:}{Integer greater than 50.
Number of Terms added in the series represantion of the density.}
}
\value{
Samples from the MPG(n, alpha, nu) distribution.
}
\description{
Generates samples from Modfified Ploya Gamma (MPG) distribution.
}
\examples{
library(Bessel)
data=rMPG(n=1000, alpha=5, nu=0)
#' data=rMPG(n=1000, alpha=50, nu=0)
}
