% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMPG_nu_Half.R
\name{rMPG_nu_Half}
\alias{rMPG_nu_Half}
\title{rMPG_nu_Half(alpha) generates random sample from the Modified Ploya Gamma distribution.}
\usage{
rMPG_nu_Half(n = 1, alpha)
}
\arguments{
\item{n:}{Number of samples to be generated.}

\item{alpha:}{The parameter of the distribution, a positive number.}
}
\value{
The sample from the MPG(nu=.5, \code{alpha}) dostribution.
}
\description{
rMPG_nu_Half(alpha) generates random sample from the Modified Ploya Gamma distribution.
}
\examples{
library(ghyp)
rMPG_nu_Half(10,1)
rMPG_nu_Half(200,100)
rMPG_nu_Half(5,10000)
}
