% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slice_DW.R
\name{VONF_2D_DW_SLICE}
\alias{VONF_2D_DW_SLICE}
\title{The functon is a implementation of the Slice sampling algorithm
to conduct Bayesian inference of circular data.}
\usage{
VONF_2D_DW_SLICE(Y, MCSamplerSize = 100, kappa_start = NULL)
}
\arguments{
\item{Y:}{n X 3 matrix containing n directional data of dimension d.
norm of each row of the matrix Y is 1.}

\item{kappa_start:}{A positive number, starting value for the MCMC algorithm.
If set NUll the Default procedure to get the initial value will be used.}

\item{MCSamplerSize:}{Number of MCMC samples required to be generated.}
}
\value{
MCMC samples from for  mode and  concentration parameter of Vonmises distribution.
}
\description{
The functon is a implementation of the Slice sampling algorithm
to conduct Bayesian inference of circular data.
}
\examples{
library(Rfast)
data = rvmf(n =10, mu=c(1,0,0),k = 10)
lst_DW=VONF_2D_DW_SLICE(Y, MCSamplerSize=1000)
}
\references{
\itemize{

 \item   [1] Damien, P., & Walker, S. (1999). A full Bayesian analysis of circular data
  using the von Mises distribution.
  The Canadian Journal of Statistics/La Revue Canadienne de Statistique, 291-298.
}
}
\author{
S. Pal,  \email{subhadip.pal@louisville.edu}
}
