% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VONF_3D_GEO_DA_POSTERIOR_V1.5.R
\name{VONF_3D_GEO_DA_POSTERIOR}
\alias{VONF_3D_GEO_DA_POSTERIOR}
\title{Fits a Data Augmentation algorithm for Vonmises Distribution in 3 dimensions.
  Generates MCMC samples from Posterior of mode and concentration parameter.}
\usage{
VONF_3D_GEO_DA_POSTERIOR(Y, MCSamplerSize = 100, kappa_start = NULL)
}
\arguments{
\item{Y:}{n X 3 matrix containing n directional data of dimension d.
norm of each row of the matrix Y is 1.}

\item{kappa_start:}{A positive number, starting value for the MCMC algorithm.
If set NUll the Default procedure to get the initial value will be used.}

\item{MCSamplerSize:}{Number of MCMC samples required to be generated.}
}
\value{
MCMC samples from for  Posterior of mode and concentration parameter of Vonmises distribution.
}
\description{
Fits a Data Augmentation algorithm for Vonmises Distribution in 3 dimensions.
  Generates MCMC samples from Posterior of mode and concentration parameter.
}
\examples{
library(Rfast)
library(benchmarkme)
data = rvmf(n =10, mu=c(1,0,0),k = 10)
MC_OBject = VONF_pD_MPG_INVCDF_DA_POSTERIOR( Y=data, MCSamplerSize=50 )
}
