% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VONF_MPG_DA_POSTERIOR_V1.5.R
\name{VONF_3D_MPG_DA_POSTERIOR}
\alias{VONF_3D_MPG_DA_POSTERIOR}
\title{Fits a Data Augmentation algorithm for Vonmises Distribution in general dimension
  Generates MCMC samples from Posterior of mode and concentration parameter.}
\usage{
VONF_3D_MPG_DA_POSTERIOR(Y, kappa_start = NULL, MCSamplerSize = 5000, K = 100)
}
\arguments{
\item{Y:}{n\times 3  matrix containing n directional data of dimension 3.
norm of each row of the matrix Y is 1.}

\item{kappa_start:}{A positive number, starting value for the MCMC algorithm.
If set NUll the Default procedure to get the initial value will be used.}

\item{MCSamplerSize:}{Number of MCMC samples required to be generated.}
}
\value{
MCMC samples from for  Posterior of mode and concentration parameter of Vonmises distribution.
}
\description{
Fits a Data Augmentation algorithm for Vonmises Distribution in general dimension
  Generates MCMC samples from Posterior of mode and concentration parameter.
}
\examples{
library(Rfast)
library(benchmarkme)
library(Bessel)
library(gsl)
data=rvmf(n =10, mu=c(1,0,0),k = 10)
MC_OBject=VONF_3D_MPG_DA_POSTERIOR(Y=data, MCSamplerSize=50)
}
